backend/pyproject.toml
[project]
name = "ragnarok-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "pydantic-settings",
    "sqlmodel",
    "psycopg2",
    "python-jose",
    "passlib[bcrypt]",
    "python-multipart",
    "celery[redis]",
    "redis",
    "httpx",
    "pdfplumber",
    "pgvector",
    "email-validator",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"